# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Cultour Project",
        "title": "Cultour API",
        "contact": {
            "name": "Cultour Development Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "paths": {
...
        "/cities": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of cities with pagination and filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "List cities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of cities to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of cities to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"created_at\"",
                        "description": "Field to sort by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"desc\"",
                        "description": "Sort order (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by province ID",
                        "name": "province_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cities retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_place_models.City"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to list cities",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new city to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Create a new city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "City Information",
                        "name": "city",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_place_models.City"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "City created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_place_models.City"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid city creation details",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/cities/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search cities by various criteria",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Search cities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Search query (name, etc.)",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"created_at\"",
                        "description": "Field to sort by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"desc\"",
                        "description": "Sort order (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cities found successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_place_models.City"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid search parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/cities/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a city's details by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Get city by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "City ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_place_models.City"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing city's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Update a city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "City ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "City Update Details",
                        "name": "city",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_place_models.City"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_place_models.City"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid city update details",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a city from the system by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Delete a city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "City ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid city ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_holycann_cultour-backend_internal_place_models.City": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "City creation time",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID for the city",
                    "type": "string"
                },
                "name": {
                    "description": "City name, example: \"Jakarta\"",
                    "type": "string",
                    "example": "Jakarta"
                },
                "province_id": {
                    "description": "ID of the province where the city is located",
                    "type": "string"
                },
                "updated_at": {
                    "description": "City last update time",
                    "type": "string"
                }
            }
        },
...
        "github_com_holycann_cultour-backend_internal_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Actual response data"
                },
                "error": {
                    "description": "Detailed error information (only populated for error responses)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ErrorDetails"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "pagination": {
                    "description": "Pagination information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.Pagination"
                        }
                    ]
                },
                "request_id": {
                    "description": "Unique request identifier for tracing",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the response (success/error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ResponseStatus"
                        }
                    ]
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "details": {
                    "description": "Detailed error description",
                    "type": "string"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ResponseStatus": {
            "type": "string",
            "enum": [
                "success",
                "error"
            ],
            "x-enum-varnames": [
                "StatusSuccess",
                "StatusError"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Insert Your JWT Token. Do NOT include \"Bearer \" prefix. Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}